Leaderboard Signoff
Changes:

! Fixes alignment with first monitor.
+ Implemented sign-off functionality.

+++ Assets/GlobalGameJam/Scripts/Scoring/Score/ScoreEntry.cs
@@ -4,6 +4,6 @@ namespace GlobalGameJam
     {
         public string Name;
         public int Score;
-        public int Time;
+        public float Time;
     }
 }
